requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
python-dateutil==2.8.2
colorama==0.4.6
